[HuuNT] fix query for student and teacher in 16.07 16.08

+++ src/main/java/cnpr/lcss/model/AdminStatisticInMonthResponse.java
@@ -14,6 +14,6 @@ public class AdminStatisticInMonthResponse {
     private int newBooking;
     private int newRegisteredInfo;
     private int newStudent;
-    private int newTeacher;
+    private long newTeacher;
 
 }
